# docs/daemon/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ADF Daemon API",
    "version": "0.1.0",
    "description": "Static OpenAPI contract for the local ADF daemon. The daemon is a headless ADF runtime for loading agents, starting and stopping them, sending chat turns, streaming events, inspecting runtime diagnostics, and resolving human-in-the-loop tasks. Every route except GET /health requires `Authorization: Bearer <token>` (the per-install token in <settings dir>/daemon-token, or ADF_DAEMON_TOKEN); requests with a disallowed Host (403 host_not_allowed), a foreign Origin or Sec-Fetch-Site cross-site/same-site (403 cross_origin) are refused, and no CORS headers are sent. Credential reads return metadata only."
  },
  "security": [{ "bearerAuth": [] }],
  "servers": [
    {
      "url": "http://127.0.0.1:7385",
      "description": "Default local daemon"
    }
  ],
...
  "paths": {
...
    "/runtime/providers": {
      "get": {
        "tags": ["Runtime"],
        "summary": "Read provider diagnostics",
        "operationId": "getRuntimeProviders",
        "responses": {
          "200": {
            "description": "Provider registrations and agent usage",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProviderDiagnostics" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Runtime"],
        "summary": "Add an API-key provider; the key goes to the daemon secret store (OS keychain or the owner's passphrase file), never the settings file, and is never returned",
        "operationId": "addRuntimeProvider",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type"],
                "properties": {
                  "type": { "type": "string", "enum": ["anthropic", "openai", "openrouter", "openai-compatible"] },
                  "name": { "type": "string" },
                  "baseUrl": { "type": "string", "description": "Required for openai-compatible" },
                  "defaultModel": { "type": "string" },
                  "preset": { "type": "string", "description": "Provider catalog key (e.g. groq)" },
                  "apiKey": { "type": "string", "description": "Required except for openai-compatible (local servers)" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "`{ provider, defaultProviderId }`; provider carries `hasApiKey`, never the key", "content": { "application/json": { "schema": { "type": "object" } } } },
          "400": { "description": "codes `bad_type`, `subscription_type`, `api_key_required`, `base_url_required`, `bad_base_url`" },
          "409": { "description": "code `secret_store_locked`: set up or unlock the owner identity first" }
        }
      }
    },
    "/runtime/providers/{id}": {
      "delete": {
        "tags": ["Runtime"],
        "summary": "Remove a provider and its stored key",
        "operationId": "removeRuntimeProvider",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "`{ removed, providers }`", "content": { "application/json": { "schema": { "type": "object" } } } },
          "404": { "description": "Unknown provider" }
        }
      }
    },
...
    "/agents/{id}": {
      "get": {
        "tags": ["Agents"],
        "summary": "Get a loaded agent reference",
        "operationId": "getAgent",
        "parameters": [{ "$ref": "#/components/parameters/AgentIdPath" }],
        "responses": {
          "200": {
            "description": "Agent reference",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AgentRef" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/agents/{id}/context": {
      "get": {
        "tags": ["Resources"],
        "summary": "One loop's context usage",
        "description": "What the next request of main (or ?loop=<name>) would carry: the executor's context breakdown split into non-overlapping categories (system, files, tools, mcp:<server>, dynamic, messages; biggest first, summing to totalTokens) with their biggest items (?items=<n>, default 12, max 200), plus the compact threshold that loop auto-compacts at and where it comes from (loop, agent, model, default). available=false when the loop has no live executor.",
        "operationId": "getAgentContext",
        "parameters": [
          { "$ref": "#/components/parameters/AgentIdPath" },
          { "$ref": "#/components/parameters/LoopQuery" },
          { "name": "items", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 0, "maximum": 200 } }
        ],
        "responses": {
          "200": { "description": "Context usage", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnyJson" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-install token from <settings dir>/daemon-token (printed by `adf daemon token`), or ADF_DAEMON_TOKEN. Local adf clients send it automatically."
      }
    },
    "parameters": {
      "AgentIdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Loaded agent id, handle, or unique name, depending on endpoint support.",
        "schema": { "type": "string" }
      },
...
      "LoopQuery": {
        "name": "loop",
        "in": "query",
        "required": false,
        "description": "Cognition loop to address. Absent = main. Unknown loops answer 404.",
        "schema": { "type": "string" }
      },
...
    },
    "responses": {
...
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    },
    "schemas": {
...
      "AnyJson": {
        "description": "Arbitrary JSON value."
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" }
        },
        "additionalProperties": true
      },
...
      "ProviderDiagnostics": {
        "type": "object",
        "required": ["providers", "agentUsage"],
        "properties": {
          "providers": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "agentUsage": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        }
      },
...
      "AgentRef": {
        "type": "object",
        "required": ["id", "filePath", "config"],
        "properties": {
          "id": { "type": "string" },
          "filePath": { "type": ["string", "null"] },
          "config": { "type": "object", "additionalProperties": true }
        }
      },
...
    }
  }
}